Updated app.py
I made it so that the create a quiz option of the website to work properly on the backend.

Basically, I stored the data in a dictionary called quizzes, which stores the ID (each trivia will have an ID, I temporarily just used an integer variable to control it.), Title, description, and trivia dictionary. The trivia dictionary will have the questions, options and correct answers.

The way new quizzes will be added is in line 63, the def create_quiz(). I added comments to help explain each step, I can further explain on the meeting tomorrow if needed.

I added some print statements as a debugging for now.