Updated app.py
When you run play_quiz.html, properly formats each question and question title, description, etc into the form we have discussed. 

Quiz Title is x.000
Quiz Description is x.001
Quiz  Time is x.002

Quiz question Text x.010
Quiz question1 option 1 is x.011 
option 2 is x.012 and so on

Also made changes to play_quiz.html so that it properly works with the database and not the dict sample quiz. (Please review this part)